Improve #454: Revert link comparison behavior to default
Can't do anything nicer in `compareValue` since that is used in equality
checks, and display should not affect equality checks.

It is possible to make a "visual comparison" function which is
independent and which allows for sorting *visually*, though it's only
purpose would be in a special DataArray call and the `SORT` operator.

I may look into implementing it at a later time. As a workaround,
probably worth implementing methods for links for extracting the display
/ original path / block ID / converting this to an exact object for
comparison.