Smart refreshing (#764)
* Fix not saving refreshEnabled flag

* Fire a refresh event after the debounce interval

Using an event avoids duplication of timing and refresh-enabling
logic, and allows setting changes to take immediate effect without
a reload or layout/mode changes.

* HTML isShown/onNodeInserted APIs require 0.13.11+

* Refactor renderers for dynamic refresh

When automatic refresh is enabled, components are re-rendered when
refreshInterval milliseconds have passed since the most recent index
change, or when the component is remounted in the DOM (e.g. due to
revealing a sidebar note tab).  Changing the refresh interval or
auto-refresh setting has immediate effect.