Fix refresh not working for React-based views
The existing code had a race condition that let it get into
a state where no further updates could take place.  This
revision ensures that the view remains subscribed so long
as the container element remains unchanged.  (It also removes the
extra initialization state and lastReload state, relying on the
fact that the refreshOperation is only triggered when a view
transitions from hidden to visible, or when the index is changed.)

This should reduce the number of React refreshes, as well as
the overhead of repeatedly subscribing (and then unsubscribing)
to the various events.